# fix: optimize the ARM function for systems with weak SIMD performance