--- a/README.md
+++ b/README.md
@@ -145,21 +145,21 @@ faster than the standard library.
 | Latin-Lipsum    |  87        | 38                        | 2.3 x           |
 | Russian-Lipsum  |  7.4       | 2.7                       | 2.7 x           |
 
-On a Neoverse V1 (Graviton 3), our validation function is 1.3 to over four times
+On a Neoverse V1 (Graviton 3), our validation function is 1.3 to over five times
 faster than the standard library.
 
 | data set      | SimdUnicode speed (GB/s) | .NET speed (GB/s) |  speed up |
 |:----------------|:-----------|:--------------------------|:-------------------|
-| Twitter.json    |  12        | 8.7                        | 1.4 x           |
-| Arabic-Lipsum   |  3.4       | 2.0                       | 1.7 x           |
-| Chinese-Lipsum  |  3.4       | 2.6                       | 1.3 x           |
-| Emoji-Lipsum    |  3.4       | 0.8                       | 4.3 x           |
-| Hebrew-Lipsum   |  3.4       | 2.0                       | 1.7 x           |
-| Hindi-Lipsum    |  3.4       | 1.6                       | 2.1 x           |
-| Japanese-Lipsum |  3.4       | 2.4                       | 1.4 x           |
-| Korean-Lipsum   |  3.4       | 1.3                       | 2.6 x           |
+| Twitter.json    |  14        | 8.7                        | 1.4 x           |
+| Arabic-Lipsum   |  4.2       | 2.0                       | 2.1 x           |
+| Chinese-Lipsum  |  4.2        | 2.6                       | 1.6 x           |
+| Emoji-Lipsum    |  4.2        | 0.8                       | 5.3 x           |
+| Hebrew-Lipsum   |  4.2        | 2.0                       | 2.1 x           |
+| Hindi-Lipsum    |  4.2        | 1.6                       | 2.6 x           |
+| Japanese-Lipsum |  4.2        | 2.4                       | 1.8 x           |
+| Korean-Lipsum   |  4.2        | 1.3                       | 3.2 x           |
 | Latin-Lipsum    |  42        | 17                        | 2.5 x           |
-| Russian-Lipsum  |  3.3       | 0.95                       | 3.5 x           |
+| Russian-Lipsum  |  4.2        | 0.95                       | 4.4 x           |
 
 
 On a Qualcomm 8cx gen3 (Windows Dev Kit 2023), we get roughly the same relative performance