To optimize on weak ARM cores, I am using a graviton 2 from AWS. It is based on Neoverse N1. The trick is to do the sum across only once per string (or, once per blocks of ~4kB for long strings). 

New results on @EgorBo's data




|                              Method |                  FileName |      Mean |    Error |   StdDev | Speed (GB/s) |
|------------------------------------ |-------------------------- |----------:|---------:|---------:|------------- |
|          SIMDUtf8ValidationRealData | data/Bogatov1069.utf8.txt | 510.69 ns | 4.810 ns | 0.264 ns |         2.09 |
|          SIMDUtf8ValidationRealData |  data/Bogatov136.utf8.txt |  92.22 ns | 0.438 ns | 0.024 ns |         1.47 |
|          SIMDUtf8ValidationRealData |  data/Bogatov286.utf8.txt | 176.81 ns | 1.814 ns | 0.099 ns |         1.62 |
|          SIMDUtf8ValidationRealData |  data/Bogatov527.utf8.txt | 284.26 ns | 1.160 ns | 0.064 ns |         1.85 |

Old results on  @EgorBo's data

|                              Method |                  FileName |      Mean |    Error |   StdDev | Speed (GB/s) |
|------------------------------------ |-------------------------- |----------:|---------:|---------:|------------- |
|          SIMDUtf8ValidationRealData | data/Bogatov1069.utf8.txt | 576.68 ns | 89.757 ns | 4.920 ns |         1.85 |
|          SIMDUtf8ValidationRealData |  data/Bogatov136.utf8.txt |  98.05 ns |  4.872 ns | 0.267 ns |         1.39 |
|          SIMDUtf8ValidationRealData |  data/Bogatov286.utf8.txt | 193.19 ns | 13.463 ns | 0.738 ns |         1.48 |
|          SIMDUtf8ValidationRealData |  data/Bogatov527.utf8.txt | 318.96 ns | 11.377 ns | 0.624 ns |         1.65 |

New results on Twitter:


|                              Method |          FileName |      Mean |    Error |   StdDev | Speed (GB/s) |
|------------------------------------ |------------------ |----------:|---------:|---------:|------------- |
|          SIMDUtf8ValidationRealData | data/twitter.json |  81.39 us | 1.671 us | 0.092 us |         7.76 |

Old results on Twitter:


|                              Method |          FileName |      Mean |    Error |   StdDev | Speed (GB/s) |
|------------------------------------ |------------------ |----------:|---------:|---------:|------------- |
|          SIMDUtf8ValidationRealData | data/twitter.json |  90.52 us | 1.111 us | 0.061 us |         6.98 |

New results on Lipsum:


|                              Method |                      FileName |       Mean |     Error |    StdDev | Speed (GB/s) |
|------------------------------------ |------------------------------ |-----------:|----------:|----------:|------------- |
|          SIMDUtf8ValidationRealData |   data/Arabic-Lipsum.utf8.txt |  33.085 us | 0.0820 us | 0.0045 us |         2.47 |
|          SIMDUtf8ValidationRealData |  data/Chinese-Lipsum.utf8.txt |  28.135 us | 0.0356 us | 0.0020 us |         2.48 |
|          SIMDUtf8ValidationRealData |    data/Emoji-Lipsum.utf8.txt |  26.412 us | 0.0459 us | 0.0025 us |         2.48 |
|          SIMDUtf8ValidationRealData |   data/Hebrew-Lipsum.utf8.txt |  26.829 us | 0.2250 us | 0.0123 us |         2.48 |
|          SIMDUtf8ValidationRealData |    data/Hindi-Lipsum.utf8.txt |  37.910 us | 0.1988 us | 0.0109 us |         2.32 |
|          SIMDUtf8ValidationRealData | data/Japanese-Lipsum.utf8.txt |  27.758 us | 9.4075 us | 0.5157 us |         2.44 |
|          SIMDUtf8ValidationRealData |   data/Korean-Lipsum.utf8.txt |  26.945 us | 1.8570 us | 0.1018 us |         2.47 |
|          SIMDUtf8ValidationRealData |    data/Latin-Lipsum.utf8.txt |   3.737 us | 0.0205 us | 0.0011 us |        23.26 |
|          SIMDUtf8ValidationRealData |  data/Russian-Lipsum.utf8.txt |  45.083 us | 0.0307 us | 0.0017 us |         2.32 |

Old results on Lipsum:


|                              Method |                      FileName |       Mean |      Error |    StdDev | Speed (GB/s) |
|------------------------------------ |------------------------------ |-----------:|-----------:|----------:|------------- |
|          SIMDUtf8ValidationRealData |   data/Arabic-Lipsum.utf8.txt |  37.712 us |  0.1409 us | 0.0077 us |         2.17 |
|          SIMDUtf8ValidationRealData |  data/Chinese-Lipsum.utf8.txt |  33.545 us |  0.9888 us | 0.0542 us |         2.08 |
|          SIMDUtf8ValidationRealData |    data/Emoji-Lipsum.utf8.txt |  37.238 us |  0.1203 us | 0.0066 us |         1.76 |
|          SIMDUtf8ValidationRealData |   data/Hebrew-Lipsum.utf8.txt |  33.957 us |  0.0790 us | 0.0043 us |         1.96 |
|          SIMDUtf8ValidationRealData |    data/Hindi-Lipsum.utf8.txt |  44.110 us |  0.0708 us | 0.0039 us |         1.99 |
|          SIMDUtf8ValidationRealData | data/Japanese-Lipsum.utf8.txt |  32.090 us |  1.5966 us | 0.0875 us |         2.11 |
|          SIMDUtf8ValidationRealData |   data/Korean-Lipsum.utf8.txt |  30.836 us |  0.3372 us | 0.0185 us |         2.16 |
|          SIMDUtf8ValidationRealData |    data/Latin-Lipsum.utf8.txt |   3.687 us |  0.1805 us | 0.0099 us |        23.58 |
|          SIMDUtf8ValidationRealData |  data/Russian-Lipsum.utf8.txt |  50.254 us |  3.7907 us | 0.2078 us |         2.08 |